+++ scripts/build.js
@@ -11,6 +11,7 @@ const CSS_FILES = [
     "css/signature.css",
     "css/text.css",
     "css/viewport.css",
+    "css/plugins/calligraphy.css",
     "css/plugins/collapsible.css",
     "css/plugins/console.css",
     "css/plugins/inspiration.css",
@@ -21,6 +22,7 @@ const CSS_FILES = [
 
 const JS_FILES = [
     "js/util.js",
+    "js/plugins/calligraphy.js",
     "js/plugins/collapsible.js",
     "js/plugins/console.js",
     "js/plugins/fonts.js",